feat: just movie list table and split components
