+++ core/components/Footer/index.tsx
@@ -1,4 +1,5 @@
-import { Box, Typography } from '@mui/material';
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
 
 // import MUI components
 import { memo } from 'react';
