+++ core/components/Modal/AuthModal.tsx
@@ -3,16 +3,16 @@ import Link from 'next/link';
 // import SWR
 import { mutate } from 'swr';
 
-// import MUI Components
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-} from '@mui/material';
+// import local service
 import localServ from '../../services/localServ';
+
+// import MUI Components
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
 
 const AuthModal = ({ open, message }: { open: boolean; message: string }) => {
   const handleLogin = () => {
