+++ core/components/Movie/List/MovieTable.tsx
@@ -1,182 +1,39 @@
-import { useState } from 'react';
-
-// import interface and type
-import {
-  InterfaceMovie,
-  InterfaceMoviePagi,
-} from '../../../interface/models/movie';
-
-// import MUI components
-import {
-  Box,
-  Checkbox,
-  FormControl,
-  FormControlLabel,
-  MenuItem,
-  Paper,
-  Select,
-  SelectChangeEvent,
-  Switch,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TablePagination,
-  TableRow,
-  TableSortLabel,
-  Toolbar,
-  Tooltip,
-  Typography,
-} from '@mui/material';
-import visuallyHidden from '@mui/utils/visuallyHidden';
-import BorderColorIcon from '@mui/icons-material/BorderColor';
-import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
+import { ChangeEvent, useState } from 'react';
 
 // import local library
 import { useSWRConfig } from 'swr';
 import moment from 'moment';
 
-function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
+// import interface and type
+import {
+  InterfaceMovie,
+  InterfaceMoviePagi,
+} from 'core/interface/models/movie';
+import { Order } from 'core/interface/common/index.interface';
 
-type Order = 'asc' | 'desc';
-function getComparator<Key extends keyof any>(
-  order: Order,
-  orderBy: Key
-): (
-  a: { [key in Key]: number | string | boolean },
-  b: { [key in Key]: number | string | boolean }
-) => number {
-  return order === 'desc'
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
+// import local components
+import EnhancedTableHead from './TableHead';
 
-function stableSort<T>(
-  array: readonly T[],
-  comparator: (a: T, b: T) => number
-) {
-  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-}
+// import local utils
+import { getComparator } from 'core/utilities';
 
-interface HeadCell {
-  disablePadding: boolean;
-  id: keyof InterfaceMovie;
-  label: string;
-  cellAlign: 'center' | 'left' | 'right' | 'inherit' | 'justify';
-}
-const headCells: readonly HeadCell[] = [
-  { id: 'maPhim', cellAlign: 'left', disablePadding: true, label: 'Mã phim' },
-  {
-    id: 'tenPhim',
-    cellAlign: 'left',
-    disablePadding: false,
-    label: 'Tên phim',
-  },
-  {
-    id: 'ngayKhoiChieu',
-    cellAlign: 'right',
-    disablePadding: false,
-    label: 'Ngày khởi chiếu',
-  },
-  {
-    id: 'dangChieu',
-    cellAlign: 'right',
-    disablePadding: false,
-    label: 'Tình trạng',
-  },
-  { id: 'hot', cellAlign: 'right', disablePadding: false, label: 'Hot' },
-  {
-    id: 'danhGia',
-    cellAlign: 'right',
-    disablePadding: false,
-    label: 'Đánh giá',
-  },
-];
-
-interface EnhancedTableProps {
-  order: Order;
-  orderBy: string;
-  onRequestSort: (
-    event: React.MouseEvent<unknown>,
-    property: keyof InterfaceMovie
-  ) => void;
-}
-
-const EnhancedTableHead = (props: EnhancedTableProps) => {
-  const { order, orderBy, onRequestSort } = props;
-  const createSortHandler =
-    (property: keyof InterfaceMovie) => (event: React.MouseEvent<unknown>) => {
-      onRequestSort(event, property);
-    };
-
-  return (
-    <TableHead>
-      <TableRow>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.cellAlign}
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component="span" sx={visuallyHidden}>
-                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-        <TableCell key="action" align="center" sortDirection={false}>
-          Action
-        </TableCell>
-      </TableRow>
-    </TableHead>
-  );
-};
-
-function TableToolbar() {
-  return (
-    <Toolbar
-      sx={{
-        pl: { sm: 2 },
-        pr: { xs: 1, sm: 1 },
-      }}
-    >
-      <Typography
-        sx={{ flex: '1 1 100%' }}
-        variant="h6"
-        id="tableTitle"
-        component="div"
-      >
-        Movie List
-      </Typography>
-    </Toolbar>
-  );
-}
+// import MUI components
+import Box from '@mui/material/Box';
+import Checkbox from '@mui/material/Checkbox';
+import FormControl from '@mui/material/FormControl';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import MenuItem from '@mui/material/MenuItem';
+import Select from '@mui/material/Select';
+import { SelectChangeEvent } from '@mui/material/Select/SelectInput';
+import Switch from '@mui/material/Switch';
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableContainer from '@mui/material/TableContainer';
+import TablePagination from '@mui/material/TablePagination';
+import TableRow from '@mui/material/TableRow';
+import BorderColorIcon from '@mui/icons-material/BorderColor';
+import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 
 const MovieTable = () => {
   const { cache } = useSWRConfig();
@@ -216,7 +73,12 @@ const MovieTable = () => {
     setDense(event.target.checked);
   };
 
-  const handleSelectStatus = (e: SelectChangeEvent) => {};
+  const handleSelectStatus = (e: SelectChangeEvent) => {
+    console.log(e);
+  };
+  const handleCheckHot = (e: ChangeEvent<HTMLInputElement>) => {
+    console.log(e.target.checked);
+  };
 
   // Avoid a layout jump when reaching the last page with empty rows.
   const emptyRows =
@@ -224,101 +86,100 @@ const MovieTable = () => {
 
   return (
     <Box sx={{ width: '100%' }}>
-      <Paper sx={{ width: '100%', mb: 2 }}>
-        <TableToolbar />
-        <TableContainer>
-          <Table
-            sx={{ minWidth: 750 }}
-            aria-labelledby="tableTitle"
-            size={dense ? 'small' : 'medium'}
-          >
-            <EnhancedTableHead
-              order={order}
-              orderBy={orderBy}
-              onRequestSort={handleRequestSort}
-            />
-            <TableBody>
-              {movieList
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .sort(getComparator(order, orderBy))
-                .map((row, index) => {
-                  const labelId = `enhanced-table-checkbox-${index}`;
+      {/* <Paper sx={{ width: '100%', mb: 2, overflow: 'hidden' }}> */}
+      <TableContainer>
+        <Table
+          aria-label="sticky table"
+          aria-labelledby="tableTitle"
+          size={dense ? 'small' : 'medium'}
+          sx={{ minWidth: 750 }}
+        >
+          <EnhancedTableHead
+            order={order}
+            orderBy={orderBy}
+            onRequestSort={handleRequestSort}
+          />
+          <TableBody>
+            {movieList
+              .sort(getComparator(order, orderBy))
+              .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+              .map((row, index) => {
+                const labelId = `enhanced-table-checkbox-${index}`;
 
-                  return (
-                    <TableRow
-                      hover
-                      role="checkbox"
-                      tabIndex={-1}
-                      key={row.maPhim}
-                    >
-                      <TableCell
-                        component="th"
-                        id={labelId}
-                        scope="row"
-                        padding="none"
-                      >
-                        {row.maPhim}
-                      </TableCell>
-                      <TableCell align="left">{row.tenPhim}</TableCell>
-                      <TableCell align="right">
-                        {moment(row.ngayKhoiChieu).format('DD/MM/YYYY')}
-                      </TableCell>
-                      <TableCell align="right">
-                        <FormControl size="small">
-                          <Select
-                            value={row.dangChieu.toString()}
-                            onChange={handleSelectStatus}
-                          >
-                            <MenuItem value="true">Đang chiếu</MenuItem>
-                            <MenuItem value="false">Sắp chiếu</MenuItem>
-                          </Select>
-                        </FormControl>
-                      </TableCell>
-                      <TableCell align="right">
-                        <Checkbox color="primary" checked={row.hot} />
-                      </TableCell>
-                      <TableCell align="right">{row.danhGia}</TableCell>
-                      <TableCell align="center">
-                        <Box
-                          component="div"
-                          sx={{
-                            display: 'flex',
-                            justifyContent: 'center',
-                            alignItems: 'center',
-                          }}
+                return (
+                  <TableRow
+                    hover
+                    role="checkbox"
+                    tabIndex={-1}
+                    key={row.maPhim}
+                  >
+                    <TableCell component="th" id={labelId} scope="row">
+                      {row.maPhim}
+                    </TableCell>
+                    <TableCell align="left">{row.tenPhim}</TableCell>
+                    <TableCell align="right">
+                      {moment(row.ngayKhoiChieu).format('DD/MM/YYYY')}
+                    </TableCell>
+                    <TableCell align="right">
+                      <FormControl size="small">
+                        <Select
+                          defaultValue={row.dangChieu.toString()}
+                          onChange={handleSelectStatus}
                         >
-                          <BorderColorIcon
-                            sx={{ mr: '0.5rem' }}
-                            color="warning"
-                          />
-                          <DeleteOutlineIcon color="error" />
-                        </Box>
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-              {emptyRows > 0 && (
-                <TableRow
-                  style={{
-                    height: (dense ? 33 : 53) * emptyRows,
-                  }}
-                >
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
-          </Table>
-        </TableContainer>
-        <TablePagination
-          rowsPerPageOptions={[5, 10]}
-          component="div"
-          count={movieList.length}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          onPageChange={handleChangePage}
-          onRowsPerPageChange={handleChangeRowsPerPage}
-        />
-      </Paper>
+                          <MenuItem value="true">Đang chiếu</MenuItem>
+                          <MenuItem value="false">Sắp chiếu</MenuItem>
+                        </Select>
+                      </FormControl>
+                    </TableCell>
+                    <TableCell align="right">
+                      <Checkbox
+                        color="primary"
+                        defaultChecked={row.hot}
+                        onChange={handleCheckHot}
+                      />
+                    </TableCell>
+                    <TableCell align="right">{row.danhGia}</TableCell>
+                    <TableCell align="center">
+                      <Box
+                        component="div"
+                        sx={{
+                          display: 'flex',
+                          justifyContent: 'center',
+                          alignItems: 'center',
+                        }}
+                      >
+                        <BorderColorIcon
+                          sx={{ mr: '0.5rem' }}
+                          color="warning"
+                        />
+                        <DeleteOutlineIcon color="error" />
+                      </Box>
+                    </TableCell>
+                  </TableRow>
+                );
+              })}
+            {emptyRows > 0 && (
+              <TableRow
+                style={{
+                  height: (dense ? 33 : 53) * emptyRows,
+                }}
+              >
+                <TableCell colSpan={7} />
+              </TableRow>
+            )}
+          </TableBody>
+        </Table>
+      </TableContainer>
+      <TablePagination
+        rowsPerPageOptions={[5, 10]}
+        component="div"
+        count={movieList.length}
+        rowsPerPage={rowsPerPage}
+        page={page}
+        onPageChange={handleChangePage}
+        onRowsPerPageChange={handleChangeRowsPerPage}
+      />
+      {/* </Paper> */}
       <FormControlLabel
         control={<Switch checked={dense} onChange={handleChangeDense} />}
         label="Dense padding"
