+++ core/interface/components/index.interface.ts
@@ -1,4 +1,3 @@
-import { Moment } from 'moment';
 import {
   Dispatch,
   MutableRefObject,
@@ -6,20 +5,28 @@ import {
   SetStateAction,
 } from 'react';
 
+// import types and interfaces
+import { Moment } from 'moment';
+import { InterfaceMovie } from '../models/movie';
+import { Order } from '../common/index.interface';
+
 export interface InterfaceBaseProps {
   className?: string;
   children?: ReactElement;
 }
 
+// Spinner
 export interface InterfaceInnerSpinner extends InterfaceBaseProps {
   size?: string;
   thickness?: number;
 }
 
+// Login
 export interface InterfaceLoginPageComponent extends InterfaceBaseProps {
   setLoading: Dispatch<SetStateAction<boolean>>;
 }
 
+// Layout
 export interface InterfaceSidebar extends InterfaceBaseProps {
   sideOpen: boolean;
   setSideOpen: Dispatch<SetStateAction<boolean>>;
@@ -27,8 +34,24 @@ export interface InterfaceSidebar extends InterfaceBaseProps {
 
 export interface InterfaceHeader extends InterfaceSidebar {}
 
+// Movie List
 export interface InterfaceSearchBar extends InterfaceBaseProps {
   tenPhimRef: MutableRefObject<HTMLInputElement | null>;
   fromDateRef: MutableRefObject<Moment | null>;
   toDateRef: MutableRefObject<Moment | null>;
+}
+
+export interface HeadCell {
+  id: keyof InterfaceMovie;
+  label: string;
+  cellAlign: 'center' | 'left' | 'right' | 'inherit' | 'justify';
+}
+
+export interface InterfaceEnhancedTableHead extends InterfaceBaseProps {
+  order: Order;
+  orderBy: string;
+  onRequestSort: (
+    event: React.MouseEvent<unknown>,
+    property: keyof InterfaceMovie
+  ) => void;
 }
