+++ pages/index.tsx
@@ -16,7 +16,9 @@ import SearchBar from '../core/components/Movie/List/SearchBar';
 import MovieTable from '../core/components/Movie/List/MovieTable';
 
 // import MUI components
-import { Box, Button, Typography } from '@mui/material';
+import Box from '@mui/material/Box';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
 
 const Home: NextPageWithLayout = () => {
   const tenPhimRef = useRef<HTMLInputElement | null>(null);
